refactor: Search to separate folder and create corresponding css file
