+++ src/components/Home.tsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import { useLocation, useNavigate } from 'react-router-dom';
-import Search from './Search';
+import Search from './Search/Search';
 import PokemonList from './PokemonList';
 import Pagination from '../components/Pagination';
 import PokemonDetails from '../components/PokemonDetails';
