+++ src/components/Search/Search.css
@@ -1,0 +1,16 @@
+.search {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.search input {
+  padding: 8px;
+  font-size: 16px;
+}
+
+.search button {
+  padding: 8px 16px;
+  font-size: 16px;
+  margin-left: 8px;
+}
