+++ src/components/Search/Search.tsx
@@ -1,5 +1,6 @@
 import React from 'react';
-import useSearchTerm from '../hooks/useSearchTerm';
+import useSearchTerm from '../../hooks/useSearchTerm';
+import './Search.css';
 
 interface SearchProps {
   onSearch: (term: string) => void;
