+++ src/index.css
@@ -9,23 +9,6 @@
   padding: 20px;
 }
 
-.search {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.search input {
-  padding: 8px;
-  font-size: 16px;
-}
-
-.search button {
-  padding: 8px 16px;
-  font-size: 16px;
-  margin-left: 8px;
-}
-
 li {
   list-style-type: none;
 }
